-Added 'Choose' 'Move' 'Do' action template.
The template is meant to be a flexible action template for any situation where a bot has to choose an target (from a list), move to this target and after arrival do some action.
The template uses a ai-value specified in the constructor to retrieve one or multiple targets. The choosing and moving is then handled by default code that can be overridden as needed.
By default the action-node relevance/priority is used to continue the movement and do required action at arrival.